+++ prj/vs2022/Neon.vcxproj
@@ -121,9 +121,10 @@
     <ClCompile Include="..\..\src\Simd\SimdNeonSynetGridSample2d32fBlZ.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdNeonSynetInnerProduct32f.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdNeonSynetInnerProduct8i.cpp" />
-    <ClCompile Include="..\..\src\Simd\SimdNeonSynetMergedConvolution32fCd.cpp" />
-    <ClCompile Include="..\..\src\Simd\SimdNeonSynetMergedConvolution32fCdc.cpp" />
-    <ClCompile Include="..\..\src\Simd\SimdNeonSynetMergedConvolution32fDc.cpp" />
+    <ClCompile Include="..\..\src\Simd\SimdNeonSynetMergedConvolution32f.cpp" />
+    <ClCompile Include="..\..\src\Simd\SimdNeonSynetMergedConvolution32fDepthwise.cpp" />
+    <ClCompile Include="..\..\src\Simd\SimdNeonSynetMergedConvolution32fInput.cpp" />
+    <ClCompile Include="..\..\src\Simd\SimdNeonSynetMergedConvolution32fOutput.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdNeonSynetPermute.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdNeonSynetPooling.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdNeonSynetPoolingMax16b.cpp" />
